fix(filid): apply cross-review corrections
- Declare boundary exemptions (direct internal imports for barrel pull-in
  avoidance, flat root-peer layout) in the five PR-touched DETAIL.md files
- Split setValueByPointer.test.ts 41->22+19 and applyPatch.test.ts 50->31+19
  to satisfy the test-record case cap without dropping coverage
